# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ThreatFocus Watchlist Service",
    "description": "Tenant-scoped watchlist and indicator management API.",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "Health",
      "description": "Service availability checks."
    },
    {
      "name": "Watchlists",
      "description": "Create and manage tenant watchlists."
    },
    {
      "name": "Indicators",
      "description": "Add and remove indicators within a watchlist."
    },
    {
      "name": "Snapshots",
      "description": "Read the effective watchlist configuration for a tenant."
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "Check service health",
        "description": "Returns the current availability status of the watchlist service.",
        "responses": {
          "200": {
            "description": "The service is healthy.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "HEALTHY"
                    },
                    "service": {
                      "type": "string",
                      "example": "watchlist-service"
                    }
                  },
                  "required": [
                    "status",
                    "service"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/watchlists": {
      "post": {
        "tags": [
          "Watchlists"
        ],
        "summary": "Create a watchlist",
        "description": "Creates a new watchlist for the authenticated tenant.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The newly created watchlist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Watchlists"
        ],
        "summary": "List watchlists",
        "description": "Returns a cursor-paginated list of watchlists for the authenticated tenant.",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "description": "The unique identifier for the tenant.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of watchlists and an optional continuation cursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedWatchlistsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/watchlists/{watchlist_id}": {
      "parameters": [
        {
          "name": "watchlist_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "tags": [
          "Watchlists"
        ],
        "summary": "Get a watchlist",
        "description": "Returns one watchlist belonging to the authenticated tenant.",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "description": "The unique identifier for the tenant.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Watchlist details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Watchlists"
        ],
        "summary": "Update a watchlist",
        "description": "Updates the supplied fields on a watchlist belonging to the authenticated tenant.",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "description": "The unique identifier for the tenant.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated watchlist details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Watchlists"
        ],
        "summary": "Delete a watchlist",
        "description": "Deletes a watchlist and its tenant-scoped configuration.",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "description": "The unique identifier for the tenant.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The watchlist was deleted successfully."
          }
        }
      }
    },
    "/v1/watchlists/{watchlist_id}/indicators": {
      "parameters": [
        {
          "name": "watchlist_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "tags": [
          "Indicators"
        ],
        "summary": "Create an indicator",
        "description": "Validates, normalizes, and adds an indicator to a watchlist.",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "description": "The unique identifier for the tenant.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndicatorCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The newly created indicator.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndicatorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Indicators"
        ],
        "summary": "List indicators",
        "description": "Returns a cursor-paginated list of indicators in a watchlist.",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "description": "The unique identifier for the tenant.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of indicators and an optional continuation cursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedIndicatorsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/watchlists/{watchlist_id}/indicators/{indicator_id}": {
      "parameters": [
        {
          "name": "watchlist_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "indicator_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "delete": {
        "tags": [
          "Indicators"
        ],
        "summary": "Delete an indicator",
        "description": "Deletes an indicator from a watchlist.",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "description": "The unique identifier for the tenant.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The indicator was deleted successfully."
          }
        }
      }
    },
    "/v1/snapshots": {
      "get": {
        "tags": [
          "Snapshots"
        ],
        "summary": "Get configuration snapshot",
        "description": "Returns the effective watchlist and indicator configuration for the authenticated tenant, optionally filtered by indicator type.",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "description": "The unique identifier for the tenant.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "indicator_types",
            "in": "query",
            "required": false,
            "description": "Optional indicator type filter.",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/IndicatorType"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The tenant's current watchlist configuration.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "IndicatorType": {
        "type": "string",
        "enum": [
          "EMAIL",
          "DOMAIN",
          "IPV4",
          "IPV6",
          "MD5",
          "SHA1",
          "SHA256",
          "KEYWORD"
        ]
      },
      "MatchType": {
        "type": "string",
        "enum": [
          "EXACT",
          "SUFFIX",
          "PREFIX",
          "CONTAINS"
        ]
      },
      "WatchlistCreate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "description": {
            "type": "string",
            "maxLength": 512,
            "nullable": true
          },
          "enabled": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "name"
        ]
      },
      "WatchlistUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128,
            "nullable": true
          },
          "description": {
            "type": "string",
            "maxLength": 512,
            "nullable": true
          },
          "enabled": {
            "type": "boolean",
            "nullable": true
          }
        },
        "required": [
          "name",
          "description",
          "enabled"
        ]
      },
      "WatchlistResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "enabled": {
            "type": "boolean"
          },
          "revision": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenantId",
          "name",
          "description",
          "enabled",
          "revision",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedWatchlistsResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WatchlistResponse"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "items"
        ]
      },
      "IndicatorCreate": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/IndicatorType"
          },
          "value": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "maxLength": 512,
            "nullable": true
          },
          "matchType": {
            "$ref": "#/components/schemas/MatchType"
          },
          "enabled": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "type",
          "value",
          "description",
          "matchType"
        ]
      },
      "IndicatorResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "watchlistId": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/IndicatorType"
          },
          "value": {
            "type": "string"
          },
          "normalizedValue": {
            "type": "string"
          },
          "matchType": {
            "$ref": "#/components/schemas/MatchType"
          },
          "enabled": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "watchlistId",
          "tenantId",
          "type",
          "value",
          "normalizedValue",
          "matchType",
          "enabled",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedIndicatorsResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IndicatorResponse"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "items"
        ]
      },
      "SnapshotIndicator": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/IndicatorType"
          },
          "normalizedValue": {
            "type": "string"
          },
          "matchType": {
            "$ref": "#/components/schemas/MatchType"
          }
        },
        "required": [
          "id",
          "type",
          "normalizedValue",
          "matchType"
        ]
      },
      "SnapshotWatchlist": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "revision": {
            "type": "integer"
          },
          "indicators": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SnapshotIndicator"
            }
          }
        },
        "required": [
          "id",
          "name",
          "revision",
          "indicators"
        ]
      },
      "SnapshotResponse": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string"
          },
          "watchlists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SnapshotWatchlist"
            }
          }
        },
        "required": [
          "tenantId",
          "watchlists"
        ]
      }
    }
  }
}